Ported to LP and tested: (ma-github-create NAME)
Function can be called from Elisp with just the repository name, and it
will either fetch the data it needs from appropriate variables, or ask
the user to provide it.  There's some tests in place for that.

The main change seems to be: instead of fetching the token from an
environment variable, now it is fetched from ~/.authinfo.gpg (or
similar).

I tried my best to be organized, but it's messy.
Next step is to reorganize the commits.
Not for /functionality/, but /learning/.